Modularize dashboard
