+++ src/content/commons/interfaces/ISentiment.ts
