+++ src/content/commons/utils/DashboardMode.ts
@@ -1,0 +1,7 @@
+const DashboardMode = {
+  Full: "full",
+  Hidden: "hidden",
+  Partial: "partial",
+}
+
+export default DashboardMode
