+++ src/content/commons/utils/MetricType.ts
@@ -1,0 +1,6 @@
+const MetricType = {
+  Sentiment: "sentiment",
+  Subjectivity: "subjectivity",
+}
+
+export default MetricType
